requirements.txt
google-generativeai
pyttsx3
SpeechRecognition
wikipedia
pyautogui
pyjokes
requests
pyaudio
